Build an Apartment
This was the second project in Udacity's VR Developer nanodegree. The purpose of this project was to design and create my own apartment based on several 3D models and placing it into the 3D world, also I have learned lighting, cameras, and basic animation. At the end, I have building a new apartment scene with nice lighting and custom animation which I test it my Cardboard headset.